Move error handlers for whole project to main
We may wish to extract them to a helper class for configuring FastAPI later, but for now this will suffice.
